# reduce sleeps